Exchange pointer offsets for memHandles.
Previously, we assumed that the pointer offsets were identical across processes,
which is often the case given the MPI model, but clearly not guaranteed.

Here we always exchange the pointer offsets (even if not using a caching allocator),
to simplify the logic and to allow mixed (caching/non-caching allocator)
environments.