Add support for char, int, long, double types to collectives_all test.
Tensor type can be specified via -type parameter.  Float is the default,
as before, and 'all' (iterate over above types) is supported.

Also, allow specifying uppersizepow for the largest size of a tensor
(i.e. 2^uppersizepow is the largest tensor used in the test).  This is
used since on an M40 I'm getting OOM issues with every parameter set
to all.